requirements.txt
streamlit
torch
transformers
pandas
numpy
matplotlib
seaborn
scikit-learn
python-dotenv
google-genai